requirements.txt

flask==3.0.0
gunicorn==21.2.0
line-bot-sdk==3.7.0

pillow==10.4.0
sxtwl==2.0.6
pygsheets==2.0.6
OpenCC==1.1.9

python-dotenv==1.1.1
cloudinary==1.44.0

cryptography==46.0.3
notion-client